Threshold boss sprite
Threshold the game screen to show only the boss (black-1) on the background (white-0).

Only the red channel is used because a hunter potion is used, making all enemies appear red, while ignoring stuff like the cursor (made green).

Blurring removes the stars (white) that would otherwise show up.

Lag is a problem when this is combined with OCR, likely needs threading but this can wait.